refactor(queries): segregated into folders
