+++ src/queries/file/deleteFile.js
