+++ src/queries/folder/deleteFolder.js
