+++ src/queries/index.js
@@ -1,13 +1,13 @@
-import CREATE_FILE from './createFile'
-import DELETE_FILE from './deleteFile'
-import RENAME_FILE from './renameFile'
-import CREATE_FOLDER from './createFolder'
-import RENAME_FOLDER from './renameFolder'
-import DELETE_FOLDER from './deleteFolder'
-import GET_FOLDER from './getFolder'
-import GET_NESTED_FOLDER from './getNestedFolders'
-import IMAGE_UPLOAD from './imageUpload'
-import OPEN_FILE from './openFile'
+import CREATE_FILE from './file/createFile'
+import DELETE_FILE from './file/deleteFile'
+import RENAME_FILE from './file/renameFile'
+import IMAGE_UPLOAD from './file/imageUpload'
+import OPEN_FILE from './file/openFile'
+import CREATE_FOLDER from './folder/createFolder'
+import RENAME_FOLDER from './folder/renameFolder'
+import DELETE_FOLDER from './folder/deleteFolder'
+import GET_FOLDER from './folder/getFolder'
+import GET_NESTED_FOLDER from './folder/getNestedFolders'
 
 export {
 	CREATE_FILE,
